+++ src/apis/PlatformOrdersApi.ts
@@ -127,16 +127,16 @@ export class PlatformOrdersApi extends runtime.BaseAPI {
       headerParameters["Idempotency-Key"] = String(requestParameters["idempotencyKey"]);
     }
 
-    if (requestParameters["affinityVersion"] != null) {
-      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
-    }
-
     if (requestParameters["affinityActorId"] != null) {
       headerParameters["Affinity-Actor-Id"] = String(requestParameters["affinityActorId"]);
     }
 
     if (requestParameters["affinityActorType"] != null) {
       headerParameters["Affinity-Actor-Type"] = String(requestParameters["affinityActorType"]);
+    }
+
+    if (requestParameters["affinityVersion"] != null) {
+      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
     }
 
     if (this.configuration && this.configuration.accessToken) {
@@ -224,16 +224,16 @@ export class PlatformOrdersApi extends runtime.BaseAPI {
 
     const headerParameters: runtime.HTTPHeaders = {};
 
-    if (requestParameters["affinityVersion"] != null) {
-      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
-    }
-
     if (requestParameters["affinityActorId"] != null) {
       headerParameters["Affinity-Actor-Id"] = String(requestParameters["affinityActorId"]);
     }
 
     if (requestParameters["affinityActorType"] != null) {
       headerParameters["Affinity-Actor-Type"] = String(requestParameters["affinityActorType"]);
+    }
+
+    if (requestParameters["affinityVersion"] != null) {
+      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
     }
 
     if (this.configuration && this.configuration.accessToken) {
@@ -332,16 +332,16 @@ export class PlatformOrdersApi extends runtime.BaseAPI {
 
     const headerParameters: runtime.HTTPHeaders = {};
 
-    if (requestParameters["affinityVersion"] != null) {
-      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
-    }
-
     if (requestParameters["affinityActorId"] != null) {
       headerParameters["Affinity-Actor-Id"] = String(requestParameters["affinityActorId"]);
     }
 
     if (requestParameters["affinityActorType"] != null) {
       headerParameters["Affinity-Actor-Type"] = String(requestParameters["affinityActorType"]);
+    }
+
+    if (requestParameters["affinityVersion"] != null) {
+      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
     }
 
     if (this.configuration && this.configuration.accessToken) {
@@ -447,16 +447,16 @@ export class PlatformOrdersApi extends runtime.BaseAPI {
 
     const headerParameters: runtime.HTTPHeaders = {};
 
-    if (requestParameters["affinityVersion"] != null) {
-      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
-    }
-
     if (requestParameters["affinityActorId"] != null) {
       headerParameters["Affinity-Actor-Id"] = String(requestParameters["affinityActorId"]);
     }
 
     if (requestParameters["affinityActorType"] != null) {
       headerParameters["Affinity-Actor-Type"] = String(requestParameters["affinityActorType"]);
+    }
+
+    if (requestParameters["affinityVersion"] != null) {
+      headerParameters["Affinity-Version"] = String(requestParameters["affinityVersion"]);
     }
 
     if (this.configuration && this.configuration.accessToken) {