+++ src/index.ts
@@ -17,6 +17,8 @@ export * from "./resources/memberships";
 export * from "./resources/orders";
 export * from "./resources/patients";
 export * from "./resources/practices";
+export * from "./resources/prescription-signing-sessions";
+export * from "./resources/prescriptions";
 export * from "./resources/provider-mappings";
 export * from "./resources/request-options";
 export * from "./resources/roles";